requirements.txt
django
psycopg2
django-webpack-loader
djangorestframework
python-decouple
djangorestframework-simplejwt
pillow